2.dev/requirements.txt
google-cloud-datastore>=1.10.0